backend/requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
google-generativeai
python-dotenv
